# Server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Minus Trash API",
    "version": "1.0.0",
    "description": "Documentazione API per Minus Trash"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Report": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titolo della segnalazione"
          },
          "description": {
            "type": "string",
            "description": "Descrizione della segnalazione"
          },
          "location": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "default": "Point"
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "number"
                },
                "description": "longitude, latitude"
              }
            }
          },
          "type": {
            "type": "string",
            "enum": [
              "abbandono",
              "cestino_pieno",
              "cestino_danneggiato",
              "altro"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "segnalato",
              "verificato",
              "in_corso",
              "risolto",
              "archiviato"
            ],
            "default": "segnalato"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "urgency": {
            "type": "string",
            "enum": [
              "bassa",
              "media",
              "alta"
            ],
            "default": "media"
          }
        }
      },
      "Bin": {
        "type": "object",
        "properties": {
          "location": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "default": "Point"
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "number"
                },
                "description": "longitude, latitude"
              }
            }
          },
          "type": {
            "type": "string",
            "enum": [
              "indifferenziato",
              "carta",
              "plastica",
              "vetro",
              "organico"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "operativo",
              "manutenzione",
              "pieno",
              "danneggiato"
            ],
            "default": "operativo"
          },
          "capacity": {
            "type": "number",
            "description": "Capacità in litri"
          },
          "fillLevel": {
            "type": "number",
            "description": "Livello di riempimento (0-100)"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "paths": {
...
    "/api/waste/reports": {
      "get": {
        "summary": "Recupera tutte le segnalazioni",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista di segnalazioni",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Report"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crea una nuova segnalazione",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Report"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Segnalazione creata con successo"
          }
        }
      }
    },
    "/api/waste/reports/{id}": {
      "get": {
        "summary": "Recupera una segnalazione per ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dettaglio della segnalazione"
          },
          "404": {
            "description": "Segnalazione non trovata"
          }
        }
      },
      "put": {
        "summary": "Aggiorna una segnalazione",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Report"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Segnalazione aggiornata"
          }
        }
      },
      "delete": {
        "summary": "Elimina una segnalazione",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Segnalazione eliminata"
          }
        }
      }
    },
    "/api/waste/reports/{id}/status": {
      "put": {
        "summary": "Aggiorna lo stato di una segnalazione",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "segnalato",
                      "verificato",
                      "in_corso",
                      "risolto",
                      "archiviato"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stato aggiornato"
          }
        }
      }
    },
    "/api/waste/bins": {
      "get": {
        "summary": "Recupera tutti i cestini",
        "responses": {
          "200": {
            "description": "Lista di cestini",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bin"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crea un nuovo cestino",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bin"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cestino creato con successo"
          }
        }
      }
    },
    "/api/waste/bins/{id}": {
      "get": {
        "summary": "Recupera un cestino per ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dettaglio del cestino"
          },
          "404": {
            "description": "Cestino non trovato"
          }
        }
      },
      "put": {
        "summary": "Aggiorna un cestino",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cestino aggiornato"
          }
        }
      },
      "delete": {
        "summary": "Elimina un cestino",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cestino eliminato"
          }
        }
      }
    },
    "/api/waste/bins/{id}/status": {
      "put": {
        "summary": "Aggiorna lo stato di un cestino",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "operativo",
                      "manutenzione",
                      "pieno",
                      "danneggiato"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stato aggiornato"
          }
        }
      }
    }
  }
}